Fix an issue with fzf lua helptags not functioning
Introduced in fzf-lua commit c838d9c8 which switched the delimiter from
a space to the M.nbsp constant in the util module. This constant is
different from my own "nbsp" setting override of '\xc2\xa0' which I had
to set in order to make the fzf-lua buffers not show garbage characters.